# src/main/java/com/skillsmanagement/entity/Person.java
package com.skillsmanagement.entity;

import java.io.Serializable;
import java.util.Date;
import java.util.Set;

import javax.persistence.CascadeType;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Index;
import javax.persistence.JoinColumn;
import javax.persistence.ManyToOne;
import javax.persistence.OneToMany;
import javax.persistence.Table;

import com.fasterxml.jackson.annotation.JsonIdentityInfo;
import com.fasterxml.jackson.annotation.JsonIgnore;
import com.fasterxml.jackson.annotation.JsonManagedReference;
import com.fasterxml.jackson.annotation.ObjectIdGenerators;

import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.ToString;

@Entity
@Data
@EqualsAndHashCode(exclude = {"PersonSkills","department"})
@ToString(exclude = {"personSkills", "department"})
@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "personId")
@Table(name = "person", indexes = {
	    @Index(name = "idx_person_email", columnList = "email"),
	    @Index(name = "idx_person_department_id", columnList = "department_id")
	})
public class Person implements Serializable {
	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long personId;


	private String firstName;
	private String lastName;
	private Date dob;
	private String email;

	@OneToMany(mappedBy = "person", cascade = CascadeType.ALL, orphanRemoval = true)
	@JsonManagedReference
	@JsonIgnore
	private Set<PersonSkills> personSkills;

	@ManyToOne
	@JoinColumn(name = "department_id")
	@JsonManagedReference
	private Department department;



}