# src/main/resources/docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Skills Management API","description":"API for managing skills","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"adminuser":[]}],"tags":[{"name":"Skills and Person API","description":"Endpoints related to create/update/read/delete person and skills information"}],"paths":{"/v1/api/skills/updateSkillToPerson":{"put":{"tags":["Skills and Person API"],"summary":"Add skill to person","description":"Add Skill info to person","operationId":"updateSkillToPerson","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePersonDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusPerson"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/person/createNewPerson":{"post":{"tags":["Skills and Person API"],"summary":"Create Person","description":"Create person info","operationId":"createPerson","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"401":{"description":"Authentication Failure","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}},"200":{"description":"Person details created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}},"409":{"description":"Person Record already exists in the system. Please try to update the information","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}},"500":{"description":"Server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/skills/organizationskills":{"get":{"tags":["Skills and Person API"],"operationId":"getOrganizationSkills","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMapStringObject"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/skills/getAllSkills":{"get":{"tags":["Skills and Person API"],"summary":"Get all Skills","description":"Retrieve a list of all skills.","operationId":"getAllSkills","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageSkill"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/skills/getAllDepartment":{"get":{"tags":["Skills and Person API"],"summary":"Get all Department","description":"Retrieve all Department","operationId":"getAllDepartment","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Department"}}}}}},"security":[{"adminuser":[]}]}},"/v1/api/skills/SkillById/{skillId}":{"get":{"tags":["Skills and Person API"],"summary":"Get skill using ID","description":"Retrieve skill info by id","operationId":"getSkillById","parameters":[{"name":"skillId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authentication Failure","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusSkill"}}}},"500":{"description":"Server error","content":{"application/json":{}}},"200":{"description":"Skill Details retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseStatus"},"example":{"statusCode":"200","statusDescription":"Skill Details retrieved","statusType":"SUCCESS"}}}},"404":{"description":"skill not found","content":{"application/json":{}}}},"security":[{"adminuser":[]}]}},"/v1/api/person/getPersonById/{pId}":{"get":{"tags":["Skills and Person API"],"summary":"Get Person using id","description":"Retrieve person info using id","operationId":"getPersonById","parameters":[{"name":"pId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authentication Failure","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusPersonDTO"}}}},"200":{"description":"Person details retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusPersonDTO"}}}},"404":{"description":"Person not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusPersonDTO"}}}},"500":{"description":"Server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusPersonDTO"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/person/getAllPersons":{"get":{"tags":["Skills and Person API"],"summary":"Get all person","description":"Retrieve all person information","operationId":"getAllPersons","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagePerson"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/skills/removeSkillFromPerson/{personId}/{skillId}":{"delete":{"tags":["Skills and Person API"],"summary":"Remove skill from person","description":"Remove skill info from person","operationId":"removeSkillFromPerson","parameters":[{"name":"personId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"skillId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}}},"security":[{"adminuser":[]}]}},"/v1/api/person/deletePerson/{personId}":{"delete":{"tags":["Skills and Person API"],"summary":"Remove person","description":"Remove person info","operationId":"deletePerson","parameters":[{"name":"personId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authentication Failure","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}},"200":{"description":"Person details retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseStatusObject"}}}},"500":{"description":"Server error","content":{"application/json":{}}},"404":{"description":"Person not found","content":{"application/json":{}}}},"security":[{"adminuser":[]}]}}},"components":{"schemas":{"PersonSkillsDTO":{"required":["skillLevel"],"type":"object","properties":{"skill":{"$ref":"#/components/schemas/SkillDTO"},"skillLevel":{"type":"string","description":"Skill levels available for a person.","example":"EXPERT","enum":["EXPERT","PRACTITIONER","WORKING","AWARENESS","BEGINNER","INTERMEDIATE","ADVANCED","EXPERT"]},"yearsOfExperience":{"type":"integer","format":"int32"}}},"SkillDTO":{"required":["skillId"],"type":"object","properties":{"skillId":{"type":"integer","description":"Skill id of the skill","format":"int64","example":1}}},"UpdatePersonDTO":{"required":["departmentId","personId"],"type":"object","properties":{"personId":{"type":"integer","format":"int64"},"firstName":{"type":"string","description":"FirstName of the person","example":"Sindhuja"},"lastName":{"type":"string","description":"LastName of the person","example":"H"},"dob":{"type":"string","description":"Date of birth must be in the format dd-MM-yyyy","example":"12-12-1987"},"departmentId":{"type":"integer","format":"int64"},"personSkills":{"type":"array","items":{"$ref":"#/components/schemas/PersonSkillsDTO"}}}},"Department":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"persons":{"type":"array","items":{"$ref":"#/components/schemas/Person"}}}},"Person":{"type":"object","properties":{"personId":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dob":{"type":"string","format":"date-time"},"email":{"type":"string"},"personSkills":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/PersonSkills"}},"department":{"$ref":"#/components/schemas/Department"}}},"PersonSkills":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"person":{"$ref":"#/components/schemas/Person"},"skill":{"$ref":"#/components/schemas/Skill"},"skillLevel":{"type":"string","description":"Skill levels available for a person.","enum":["EXPERT","PRACTITIONER","WORKING","AWARENESS"]},"yearsOfExperience":{"type":"integer","format":"int32"}}},"ResponseStatusPerson":{"type":"object","properties":{"statusCode":{"type":"string"},"statusDescription":{"type":"string"},"statusType":{"type":"string"},"data":{"$ref":"#/components/schemas/Person"}}},"Skill":{"type":"object","properties":{"skillId":{"type":"integer","format":"int64"},"skillName":{"type":"string"},"personSkills":{"type":"array","items":{"$ref":"#/components/schemas/PersonSkills"}}}},"PersonDTO":{"required":["departmentId","dob","firstName","lastName"],"type":"object","properties":{"personId":{"type":"integer","format":"int64"},"firstName":{"type":"string","description":"FirstName of the person","example":"Sindhuja"},"lastName":{"type":"string","description":"LastName of the person","example":"H"},"dob":{"type":"string","description":"Date of birth must be in the format dd-MM-yyyy","example":"12-12-1987"},"email":{"type":"string","description":"The email address of the person","example":"xxx@gmail.com"},"departmentId":{"type":"integer","description":"Department id of the person","format":"int64","example":1},"personSkills":{"type":"array","items":{"$ref":"#/components/schemas/PersonSkillsDTO"}}}},"ResponseStatusObject":{"type":"object","properties":{"statusCode":{"type":"string"},"statusDescription":{"type":"string"},"statusType":{"type":"string"},"data":{"type":"object"}}},"PageMapStringObject":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageSkill":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Skill"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"ResponseStatus":{"type":"object","properties":{"statusCode":{"type":"string"},"statusDescription":{"type":"string"},"statusType":{"type":"string"},"data":{"type":"object"}}},"ResponseStatusSkill":{"type":"object","properties":{"statusCode":{"type":"string"},"statusDescription":{"type":"string"},"statusType":{"type":"string"},"data":{"$ref":"#/components/schemas/Skill"}}},"ResponseStatusPersonDTO":{"type":"object","properties":{"statusCode":{"type":"string"},"statusDescription":{"type":"string"},"statusType":{"type":"string"},"data":{"$ref":"#/components/schemas/PersonDTO"}}},"PagePerson":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Person"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}}},"securitySchemes":{"adminuser":{"type":"http","name":"API-KEY","in":"header","scheme":"basic"}}}}